pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "die-vfm"
version = "0.1.0"
description = "Die-level visual foundation model platform."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Emerging System Lab" }
]
dependencies = [
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "torch>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pylint>=3.0",
  "mypy>=1.8",
  "ruff>=0.6",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["die_vfm*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 80
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true